Added queries for demo, which was postponed from 1/13 to 1/18.
The recuriter asked me to create mock data for the data model and to demostrate the six queries in an interview on 1/13. Unfortunately, the department had a scheduling conflict, so I am uploading the revised queries and a sample of running them for reference. We are rescheduling the interview for 1/18.